# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ClinicFlow API Documentation",
        "description": "Comprehensive API documentation for ClinicFlow Backend",
        "contact": {
            "email": "admin@clinicflow.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        },
        {
            "url": "https://api.clinicflow.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/website/clinics": {
            "get": {
                "tags": [
                    "Website"
                ],
                "summary": "Get all clinics (public)",
                "operationId": "9af8a8dc4e38678d252c14dbc643ba36",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clinics list"
                    }
                }
            }
        },
        "/website/clinics/{id}": {
            "get": {
                "tags": [
                    "Website"
                ],
                "summary": "Get clinic details",
                "operationId": "4cb0f210e55c63843bd2eb3865e3b668",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clinic details"
                    }
                }
            }
        },
        "/website/clinics/{id}/services": {
            "get": {
                "tags": [
                    "Website"
                ],
                "summary": "Get clinic services",
                "operationId": "f3ef390db4d86fce35cee640416129b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Services list"
                    }
                }
            }
        },
        "/website/clinics/{id}/doctors": {
            "get": {
                "tags": [
                    "Website"
                ],
                "summary": "Get clinic doctors",
                "operationId": "36c1c2a2aef978a557b629b6040c41f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctors list"
                    }
                }
            }
        },
        "/website/services": {
            "get": {
                "tags": [
                    "Website"
                ],
                "summary": "Get all services (public)",
                "operationId": "8dac7600d33d9119c0206f0cc149023b",
                "parameters": [
                    {
                        "name": "clinic_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Services list"
                    }
                }
            }
        },
        "/website/services/{id}": {
            "get": {
                "tags": [
                    "Website"
                ],
                "summary": "Get service details",
                "operationId": "2012d0d64bfd425d677e96ed955cf8dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service details"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}